% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_filter_ui.R
\name{build_filter_ui}
\alias{build_filter_ui}
\title{Build Shiny Filter UI}
\usage{
build_filter_ui(df)
}
\arguments{
\item{df}{Data frame.}
}
\value{
A list of shiny UI elements (to be wrapped in tagList()).
}
\description{
Generate filter controls dynamically for a dataset.
\itemize{
\item Date/POSIXct columns get a date range input
\item Numeric get a slider
\item Character/factor/logical get a multi-select (unless too many uniques)
Columns named ".row_uid" are skipped.
}
}
\keyword{internal}
