% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cherry_picker_app.R
\name{cherry_picker_app}
\alias{cherry_picker_app}
\title{Core Cherry Picker App}
\usage{
cherry_picker_app(preloaded_data = NULL)
}
\arguments{
\item{preloaded_data}{Optional data frame to skip file upload.}
}
\value{
A Shiny app object.
}
\description{
Launches the Shiny app with either uploaded or preloaded data.
}
\keyword{internal}
