% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cherry_picker_ui.R
\name{cherry_picker_ui}
\alias{cherry_picker_ui}
\title{Cherry Picker UI}
\usage{
cherry_picker_ui()
}
\value{
A Shiny UI definition.
}
\description{
Internal function that builds the UI for the Cherry Picker app.
}
\keyword{internal}
