% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cherry_picker.R
\name{run_cherry_picker}
\alias{run_cherry_picker}
\title{Cherry Picker (Shiny Upload Mode)}
\usage{
run_cherry_picker()
}
\value{
A running Shiny app.
}
\description{
Launch the Cherry Picker Shiny app in hosted mode. In this mode the app
does not preload data; instead the user uploads a file (CSV) and can apply
filters before exploring it. This is the version designed for deployment
on RStudio Connect or other Shiny hosting services.
}
\examples{
\dontrun{
RunCherryPicker()
}
}
